Rename some files Refine variables.
I just did a test run and I can get a Rails app working my script flow.

After the fact, I did add LazyVim to the container:
1. AppImage from LazyVim is needed (wget tar, extract, add to path).
1. There are some dependencies I need to add to the Containerfile, most
   of which you see from `:LazyHealth`.
1. I'm going to commit this container so that I don't lose the progress.

It seems like a lot of work to figure out initially but if you develop a
repeatable system, it could be worth it!

As a solo-dev on personal projects, this all might be overkill but maybe
it'll help with the deploys; everything I'm having to learn regarding
Containers and networking should be helpful when I have to deploy and
get this running live.

I also now would love to put Neovim on my VPS.