Clean up variable referenes Update Debian image to Bookwork (Debian 12); this also helped fix an issue with installing Neovim where the GLIBC was outdated.
Docs seem to be correct. I can build a Rails new app and have Node and
Neovim installed on the system with a few scripts and very minimal
manual config from the user (only in config/database.yaml for Rails).

Pending:
- Consider putting the Neovim install script and configs in their own
  repo; as I use Neovim, I may want to push my configs up and make sure
they stay up to date when I use this.
- Neovim needs to be added to $PATH
- Don't overcustomize too much; if anyone else uses this, it should be
  simple for them to bring over their Neovim configs, whether it's lazy
or something else.
- Script to easily move the progress of this app into its own repo.

Other thoughts:
- Perhaps it should be enough to complete this project with providing
  containers that run:
  - Postgres
  - Rails:
    - with connection to Postgres
    - with base Neovim LazyVim installed

At that point, once someone talks this starter, creates a project, and
puts that in its own repo, they are free to mod the Containerfile and
scripts however they'd like.