Update Readme with new steps for creating a named volume Update postgres setup script to include the named volume
I ran through all the steps and was able to create a new rails app with
the volume. I haven't tested writing to the DB so I will hold off on
merging; when I come back to this, I'll make a simple model and test
that the DB is working.